+++ Changes.md
@@ -2,6 +2,10 @@
 
 ## v0.1
 
+- [Parser] Modifiers ( extern, volatile etc... )
+- [Parser] Inline Assembly
+- [Parser] Return Statement
+- [Parser] Variable Definitions
 - [Parser] Function Definitions
 - [Parser] Import Statements
 - [Lexer] Finished making it